Data Connectors/GoogleSecOpsDetectionAlerts/requirements.txt
azure-functions==1.17.0
azure-storage-file-share==12.15.0
httpx==0.27.2
requests==2.32.3
PyJWT==2.9.0
cryptography==43.0.1
